Populate sections in course select list
Using the sections data retrieved in commit
a9dc734c550d9966911d1c7078aec271b393f913, the courses list is now
populated with sections as appropriate.

Course offerings with only a single section (or no sections defined)
continue to be listed as is.

Course offerings with more than one section are outputted as an
`optgroup`, with each section presented as an `option`. For now, the
section’s course code is fudged by combining the containing course
offering’s course code with the section’s name. There doesn’t seem to
be an obvious way of getting a section’s code.