Improve section code generation
Improved the way section codes are generated by first checking if the
course offering’s code is in the standard Blackboard style course code
format. If it is, then proceed with replacing the course catalog type
and number from the section’s name. If not, then simply combine the
offering code with the section name as the section code.