Don't load more than 4 batches of enrollments
Administrators and other users with broad access in the D2L Brightspace
learning environment have access to a very large number of course
offerings. Limit the number of data sets loaded from
/d2l/api/lp/1.4/enrollments/myenrollments/ to 4 to avoid loading an
excessive number of courses.

This number may need adjusting in the future for actual instructors if
they teach a large number of courses or have a large list of past
courses. Changing the sort ordering should also be considered if this
comes up.